+++ neosvr_headless_webui/static/js/session.js
@@ -448,3 +448,37 @@ $('#banModal').on('show.bs.modal', function (event) {
   modal.find('.modal-body #banModalUsernameDisplay').text(username)
   modal.find('.modal-body #banModalUsername').val(username)
 });
+
+// Restart client modal
+
+$("#restartClientModalForm").submit(function(event) {
+  event.preventDefault();
+
+  var $form = $(this),
+    url = $form.attr("action");
+
+  var posting = $.post(url);
+
+  $("#restartClientModal").modal("hide");
+
+  posting.done(function(data) {
+    window.location.href = "/dashboard";
+  });
+});
+
+// Kill client modal
+
+$("#killClientModalForm").submit(function(event) {
+  event.preventDefault();
+
+  var $form = $(this),
+    url = $form.attr("action");
+
+  var posting = $.post(url);
+
+  $("#killClientModal").modal("hide");
+
+  posting.done(function(data) {
+    window.location.href = "/dashboard";
+  });
+});